Open a stdin channel to utunuds, close on exit
Relies on changes to utunuds. We were previously calling process.terminate() on utunuds in a try/catch block.  This will fail if utunuds is suid-root and we are not running as root. (i.e. it should always fail).  utunuds has been modified to look for EOF on stdin and exit gracefully when this happens.