decode IPV6 enough to read whole packets, exit on utunuds closure
Previously our rsd tunnels would fail randomly (for example, when executing launchApplication).  The cause was that we were not grabbing whole IPV6 frames to pass along to quic, but rather just reading arbirary bytes out of our tunnel and passing them along.  By chance, sometimes we'd get whole frames on clean frame boundaries and everything would work, but the unix domain socket in the middle doesn't have any notion of delivering "whole packets".  The main problem appeared to be when we exceeded the 1500 byte MTU by reading more-than-one IPV6 packet from the unix domain socket and passing it to quic as a single datagram.  But it seems there were also opportunities to mangle the data stream if we forwarded arbitrary chunks of IPV6 packets (such as keepalive 'ping' data getting interleaved with our data stream).

The utunuds service doesn't generally fail, but it was the case that if someone ran "kill" to terminate it, a process using this library would never terminate.  I've "fixed" this by generating a SIGINT when this happens, but really the code using the tunnel needs to detect that the tunnel connections have closed and exit on its own.  Then the SIGINT can be removed, but there are other priorities at the moment.