Fix writer-initialization race condition in external_utun.up()
up() was returning without a guarantee that the UDS writer had been initialized.  It was possible that the caller would start sending data that we were unable to forward, losing data and triggering error messages.  In practice, we just needed to wait a fraction of a second in the right spot for asynchronous UDS callbacks to have a chance to execute.